Optimized the code for switching map lights on/off
Instead of checking every tick and using a bunch of random timers and other things the lights in the map are randomly divided into 6 waves, There are now sunset and sunrise hooks that run which trigger the lights switching. The lights will switch according to their wave with a delay of 5 seconds between waves. This gives the look of them all not turning off at once without the cost.